Moved DataProviders under Providers folder
