+++ src/Providers/DataProviders/SqlDataProvider/Scripts/00.00.01.Uninstall.sql
