+++ src/Providers/DataProviders/SqlDataProvider/Scripts/00.00.02.Uninstall.sql
