+++ src/Providers/DataProviders/SqlDataProvider/Scripts/00.00.04.Uninstall.sql
