+++ src/Providers/DataProviders/SqlDataProvider/Scripts/00.00.05.Uninstall.sql
