+++ src/Providers/DataProviders/SqlDataProvider/SqlDataProvider.ts
@@ -1,17 +1,17 @@
 import { inject, injectable } from "inversify";
 import * as Sql from "mssql";
-import { IConfiguration } from "../../Common/Configuration";
-import { ErrorCodes, HazzatApplicationError } from "../../Common/Errors";
-import { Language } from "../../Common/Types/Language";
-import { Log } from "../../Common/Utils/Logger";
-import { SqlHelpers } from "../../Common/Utils/SqlHelpers";
-import { IFormatInfo } from "../../Models/IFormatInfo";
-import { IHymnInfo } from "../../Models/IHymnInfo";
-import { ISeasonInfo } from "../../Models/ISeasonInfo";
-import { IServiceInfo } from "../../Models/IServiceInfo";
-import { ContentType, IHazzatContent, IHymnContent, IInformationContent, ITextContent, IVariationInfo, IVerticalHazzatContent, IVideoContent, TextColumn, TextParagraph } from "../../Models/IVariationInfo";
-import { ResourceTypes } from "../../Routes/ResourceTypes";
-import { TYPES } from "../../types";
+import { IConfiguration } from "../../../Common/Configuration";
+import { ErrorCodes, HazzatApplicationError } from "../../../Common/Errors";
+import { Language } from "../../../Common/Types/Language";
+import { Log } from "../../../Common/Utils/Logger";
+import { SqlHelpers } from "../../../Common/Utils/SqlHelpers";
+import { IFormatInfo } from "../../../Models/IFormatInfo";
+import { IHymnInfo } from "../../../Models/IHymnInfo";
+import { ISeasonInfo } from "../../../Models/ISeasonInfo";
+import { IServiceInfo } from "../../../Models/IServiceInfo";
+import { ContentType, IHazzatContent, IHymnContent, IInformationContent, ITextContent, IVariationInfo, IVerticalHazzatContent, IVideoContent, TextColumn, TextParagraph } from "../../../Models/IVariationInfo";
+import { ResourceTypes } from "../../../Routes/ResourceTypes";
+import { TYPES } from "../../../types";
 import { IDataProvider } from "../IDataProvider";
 import { HazzatDbSchema } from "./HazzatDbSchema";
 import { Constants } from "./SqlConstants";
