+++ src/hazzat-api.njsproj
@@ -42,7 +42,7 @@
     <TypeScriptCompile Include="Common\Types\StringMap.ts" />
     <TypeScriptCompile Include="Common\Utils\Logger.ts" />
     <TypeScriptCompile Include="Common\Utils\SqlHelpers.ts" />
-    <TypeScriptCompile Include="DataProviders\SqlDataProvider\SqlConstants.ts" />
+    <TypeScriptCompile Include="Providers\DataProviders\SqlDataProvider\SqlConstants.ts" />
     <TypeScriptCompile Include="inversify.config.ts" />
     <TypeScriptCompile Include="Models\IVariationInfo.ts" />
     <TypeScriptCompile Include="Models\IHymnInfo.ts" />
@@ -55,24 +55,24 @@
     <TypeScriptCompile Include="types.ts" />
     <TypeScriptCompile Include="app.ts" />
     <TypeScriptCompile Include="Common\Configuration.ts" />
-    <TypeScriptCompile Include="DataProviders\SqlDataProvider\HazzatDbSchema.ts" />
+    <TypeScriptCompile Include="Providers\DataProviders\SqlDataProvider\HazzatDbSchema.ts" />
     <TypeScriptCompile Include="Models\ISeasonInfo.ts" />
-    <TypeScriptCompile Include="DataProviders\IDataProvider.ts" />
-    <TypeScriptCompile Include="DataProviders\SqlDataProvider\SqlDataProvider.ts" />
+    <TypeScriptCompile Include="Providers\DataProviders\IDataProvider.ts" />
+    <TypeScriptCompile Include="Providers\DataProviders\SqlDataProvider\SqlDataProvider.ts" />
     <TypeScriptCompile Include="Routes\HomeController.ts" />
     <TypeScriptCompile Include="Routes\SeasonsController.ts" />
     <Content Include="Configs\prod.json" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.05.Uninstall.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.05.Install.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.04.Uninstall.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.04.Install.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.03.Uninstall.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.03.Install.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.02.Uninstall.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.02.Install.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.01.Install.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\00.00.01.Uninstall.sql" />
-    <Content Include="DataProviders\SqlDataProvider\Scripts\SqlMetadata.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.05.Uninstall.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.05.Install.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.04.Uninstall.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.04.Install.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.03.Uninstall.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.03.Install.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.02.Uninstall.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.02.Install.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.01.Install.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\00.00.01.Uninstall.sql" />
+    <Content Include="Providers\DataProviders\SqlDataProvider\Scripts\SqlMetadata.sql" />
     <Content Include="tslint.json" />
     <Content Include="Configs\dev.json" />
     <Content Include="tsconfig.json" />
@@ -85,11 +85,12 @@
     <Folder Include="Common\" />
     <Folder Include="Common\Types\" />
     <Folder Include="Common\Utils\" />
-    <Folder Include="DataProviders\SqlDataProvider\Scripts\" />
+    <Folder Include="Providers\DataProviders\SqlDataProvider\Scripts\" />
     <Folder Include="Models\" />
-    <Folder Include="DataProviders\" />
-    <Folder Include="DataProviders\SqlDataProvider\" />
+    <Folder Include="Providers\DataProviders\" />
+    <Folder Include="Providers\DataProviders\SqlDataProvider\" />
     <Folder Include="Configs\" />
+    <Folder Include="Providers\" />
     <Folder Include="test\" />
     <Folder Include="Routes\" />
     <Folder Include="test\Helpers\" />
