+++ src/inversify.config.ts
@@ -1,7 +1,7 @@
 import { Container } from "inversify";
 import { Configuration, IConfiguration } from "./Common/Configuration";
-import { IDataProvider } from "./DataProviders/IDataProvider";
-import { SqlDataProvider } from "./DataProviders/SqlDataProvider/SqlDataProvider";
+import { IDataProvider } from "./Providers/DataProviders/IDataProvider";
+import { SqlDataProvider } from "./Providers/DataProviders/SqlDataProvider/SqlDataProvider";
 import { TYPES } from "./types";
 
 const myContainer = new Container();
